Separated MetaMachineExtensions into 2 new files.
